TMLE procedure for Biomarker Identication from Exposure

This function performs influence curve-based estimation of the effect of an exposure on biological expression values associated with a given biomarker, controlling for a user-specified set of baseline covariates

biomarkerTMLE_exposure(Y, W, A, a, subj_ids = NULL, family = "gaussian",
  g_lib, Q_lib)

Arguments

Y
(numeric vector) - a vector of expression values for a single biomarker.

W
(numeric matrix) - a matrix of baseline covariates to be controlled in the estimation process.

A
(numeric vector) - a discretized exposure vector (e.g., from a design matrix whose effect on expression values is of interest.

a
(numeric vector) - the levels of A against which comparisons are to be made.

subj_ids
(numeric vector) - subject IDs to be passed directly to the same subject should have the exact same numerical identifier. coerced to numeric if not provided in the appropriate form.

family
(character) - specification of error family: "binomial" or "gaussian"

g_lib
(char vector) - library of learning algorithms to be used in fitting the "g" step of the standard TMLE procedure.

Q_lib
(char vector) - library of learning algorithms to be used in fitting the "Q" step of the standard TMLE procedure.

Value

TMLE-based estimate of the relationship between biomarker expression and changes in an exposure variable, computed iteratively and saved in the tmleOut slot in a biotmle object.