TMLE procedure for Biomarker Identication from Outcome

This function performs influence curve-based estimation of the effect of an exposure on (gene) expression array values associated with a given biomarker, controlling for a user-specified set of baseline covariates

biomarkerTMLE_outcome(Y, W, A, a, family = "binomial", g_lib, Q_lib)

Arguments

Y

(numeric vector) - a vector of array expression values for a single gene

W

(numeric matrix) - a matrix of covariates to be controlled in estimation

A

(numeric vector) - a discretized exposure vector whose effect on expression gene expression values is of interest

a

(numeric vector) - the levels of A against which comparisons are to be made

family

(character) - specification of error family: "binomial" or "gaussian"

g_lib

(char vector) - library of learning algorithms to be used in ...

Q_lib

(char vector) - library of learning algorithms to be used in ...