TMLE procedure using ATE for Biomarker Identication from Exposure

This function performs influence curve-based estimation of the effect of an exposure on biological expression values associated with a given biomarker, controlling for a user-specified set of baseline covariates.

exp_biomarkertmle(Y, A, W, g_lib, Q_lib, cv_folds, ...)

Arguments

Y
A numeric vector of expression values for a given biomarker.

A
A numeric vector of discretized exposure vector (e.g., from a design matrix whose effect on expression values is of interest.

W
A Matrix of numeric values corresponding to baseline covariates to be marginalized over in the estimation process.

g_lib
A character vector identifying the library of learning algorithms to be used in fitting the propensity score P[A = a | W].

Q_lib
A character vector identifying the library of learning algorithms to be used in fitting the outcome regression E[Y | A, W].

cv_folds
A numeric scalar indicating how many folds to use in performing targeted minimum loss estimation. Cross-validated estimates are more robust, allowing relaxing of theoretical conditions and construction of conservative variance estimates.

...
Additional arguments passed to drtmle in computing the targeted minimum loss estimator of the average treatment effect.

Value

TMLE-based estimate of the relationship between biomarker expression and changes in an exposure variable, computed iteratively and saved in the tmleOut slot in a biotmle object.