The Moderated T-Statistic for Asymptotically Linear Parameters

Performs variance shrinkage via empirical Bayes procedures on influence curve transforms of the Average Treatment Effect of gene expression changes on the outcome of interest

limmatmle(biotmle, designMat, ...)

Arguments

biotmle

biotmle object as generated by biomarkertmle

designMat

a design matrix providing the contrasts to be used in the linear model fitting procedure of limma::lmFit

...

additional arguments to be passed to functions from limma

Value

biotmle object containing output from limma::lmFit and limma::topTable

Examples

library(dplyr)
data(illuminaData)
data(biomarkertmleOut)
"%ni%" = Negate("%in%")

W <- illuminaData %>%
 dplyr::select(which(colnames(.) %in% c("age", "sex", "smoking"))) %>%
 dplyr::mutate(
   age = as.numeric((age > quantile(age, 0.25))),
   sex = I(sex),
   smoking = I(smoking)
 )

A <- illuminaData %>%
 dplyr::select(which(colnames(.) %in% c("benzene")))
A <- A[, 1]

Y <- illuminaData %>%
 dplyr::select(which(colnames(.) %ni% c("age", "sex", "smoking", "benzene",
                                        "id")))
geneIDs <- colnames(Y)

design <- as.data.frame(cbind(rep(1, nrow(Y)),
                              as.numeric(A == max(unique(A)))))
colnames(design) <- c("intercept", "Tx")

limmaTMLEout <- limmatmle(biotmle = biomarkerTMLEout, IDs = NULL,
                          designMat = design)