Moderated Statistical Tests for Influence Curves

Performs variance shrinkage via the empirical Bayes procedure of LIMMA on the observed data after a transformation moving the data to influence curve space, based on the average treatment effect parameter.

modtest_ic(biotmle, design, ...)

Arguments

biotmle

biotmle object as generated by biomarkertmle

design

a design matrix providing the contrasts to be used in the linear model fitting procedure of limma::lmFit

...

additional arguments to be passed to functions from limma

Value

biotmle object containing output from limma::lmFit and limma::topTable

Examples

library(dplyr)
library(biotmleData)
library(SummarizedExperiment)
data(illuminaData)
data(biomarkertmleOut)
"%ni%" = Negate("%in%")

colData(illuminaData) <- colData(illuminaData) %>%
     data.frame %>%
     dplyr::mutate(age = as.numeric(age > median(age))) %>%
     DataFrame

varInt_index <- which(names(colData(illuminaData)) %in% "benzene")
designVar <- as.data.frame(colData(illuminaData))[, varInt_index]
design <- as.numeric(designVar == max(designVar))

limmaTMLEout <- modtest_ic(biotmle = biomarkerTMLEout, design = design)