Biomarker Discovery via Targeted Minimum Loss-Based Estimation (TMLE)

computes the (rather complicated) parameter defined as the difference in the blips in Tx effects (difference between counterfactual max, min Tx effects)

biomarkertmle(Y, W, A, type, parallel = TRUE, family = "gaussian",
  g_lib = c("SL.glmnet", "SL.randomForest", "SL.nnet", "SL.polymars",
  "SL.mean"), Q_lib = c("SL.glmnet", "SL.randomForest", "SL.nnet", "SL.mean"))

Arguments

Y

(numeric vector) - a vector of array expression values for a single gene

W

(numeric matrix) - a matrix of covariates to be controlled in estimation

A

(numeric vector) - a discretized exposure vector whose effect on expression gene expression values is of interest

type

(character) - choice of the type of TMLE to perform: "exposure" to identify biomarkers related to an exposure (input as A), or "outcome" to identify biomarkers related to an outcome (input as Y).

parallel

(logical, numeric) - whether to use or number of cores to be used when the TMLE-based estimation procedure is parallelized

family

(character) - specification of error family: "binomial" or "gaussian"

g_lib

(char vector) - library of learning algorithms to be used in ...

Q_lib

(char vector) - library of learning algorithms to be used in ...