Biomarker Evaluation with Targeted Minimum Loss-Based Estimation (TMLE)

Computes the causal target parameter defined as the difference between the biomarker expression values under treatment and those same values under no treatment, using Targeted Minimum Loss-Based Estimation.

biomarkertmle(se, varInt, ngscounts = FALSE, parallel = TRUE,
  family = "gaussian", subj_ids = NULL, g_lib = c("SL.glm",
  "SL.randomForest", "SL.nnet", "SL.polymars", "SL.mean"), Q_lib = c("SL.glm",
  "SL.randomForest", "SL.nnet", "SL.mean"))

Arguments

se
(SummarizedExperiment) - containing expression or next-generation sequencing data in the "assays" slot and a matrix of phenotype-level data in the "colData" slot.

varInt
(numeric) - indicating the column of the design matrix corresponding to the treatment or outcome of interest (in the "colData" slot of the "se" argument above).

ngscounts
(logical) - whether the data are counts generated from a next-generation sequencing (NGS) experiment (e.g., RNA-seq). The default setting assumes continuous expression measures as generated by microarray-type platforms.

parallel
(logical, numeric) - whether to use or the number of cores to be used when the TMLE-based estimation procedure is parallelized.

family
(character) - specification of error family: "binomial" or "gaussian".

subj_ids
(numeric vector) - subject IDs to be passed directly to the same subject should have the exact same numerical identifier; coerced to numeric if not provided in the appropriate form.

g_lib
(char vector) - library of learning algorithms to be used in fitting the "g" step of the standard TMLE procedure.

Q_lib
(char vector) - library of learning algorithms to be used in fitting the "Q" step of the standard TMLE procedure.

Value

S4 object of class biotmle, generated by sub-classing SummarizedExperiment, with additional slots containing tmleOut and call, among others, containing TMLE-based estimates of the relationship between a biomarker and exposure or outcome variable and the original call to this function (for user reference), respectively.

Examples

library(dplyr)
library(biotmleData)
data(illuminaData)
library(SummarizedExperiment)

#> Loading required package: GenomicRanges

#> Loading required package: stats4

#> Loading required package: BiocGenerics

#> Loading required package: parallel

#> 
#> Attaching package: ‘BiocGenerics’

#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from ‘package:dplyr’:
#> 
#>     combine, intersect, setdiff, union

#> The following objects are masked from ‘package:base’:
#> 
#>     anyDuplicated, append, as.data.frame, cbind, colMeans, colnames,
#>     colSums, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, lengths, Map, mapply, match,
#>     mget, order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rowMeans, rownames, rowSums, sapply, setdiff, sort,
#>     table, tapply, union, unique, unsplit, which, which.max, which.min

#> Loading required package: S4Vectors

#> 
#> Attaching package: ‘S4Vectors’

#> The following objects are masked from ‘package:dplyr’:
#> 
#>     first, rename

#> The following object is masked from ‘package:tidyr’:
#> 
#>     expand

#> The following object is masked from ‘package:base’:
#> 
#>     expand.grid

#> Loading required package: IRanges

#> 
#> Attaching package: ‘IRanges’

#> The following objects are masked from ‘package:dplyr’:
#> 
#>     collapse, desc, slice

#> The following object is masked from ‘package:purrr’:
#> 
#>     reduce

#> Loading required package: GenomeInfoDb

#> Loading required package: Biobase

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

#> Loading required package: DelayedArray

#> Loading required package: matrixStats

#> 
#> Attaching package: ‘matrixStats’

#> The following objects are masked from ‘package:Biobase’:
#> 
#>     anyMissing, rowMedians

#> The following object is masked from ‘package:dplyr’:
#> 
#>     count

#> 
#> Attaching package: ‘DelayedArray’

#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colMaxs, colMins, colRanges, rowMaxs, rowMins, rowRanges

#> The following object is masked from ‘package:base’:
#> 
#>     apply

"%ni%" = Negate("%in%")

colData(illuminaData) <- colData(illuminaData) %>%
     data.frame %>%
     dplyr::mutate(age = as.numeric(age > median(age))) %>%
     DataFrame

varInt_index <- which(names(colData(illuminaData)) %in% "benzene")

biomarkerTMLEout <- biomarkertmle(se = illuminaData[1, ],
                                  varInt = varInt_index,
                                  parallel = 1,
                                  family = "gaussian",
                                  g_lib = c("SL.mean"),
                                  Q_lib = c("SL.mean")
                                 )

#> [1] "Estimating target parameter for 1 of 1"

#> Loading required package: nnls