TMLE procedure using ATE for Biomarker Identication from Exposure

This function performs influence curve-based estimation of the effect of an exposure on biological expression values associated with a given biomarker, controlling for a user-specified set of baseline covariates.

biomarkerTMLE_exposure(Y, W, A, a, subj_ids = NULL,
  family = "gaussian", g_lib, Q_lib, ...)

Arguments

Y
A numeric vector of expression values for a single biomarker.

W
A Matrix of numeric values corresponding to baseline covariates to be marginalized over in the estimation process.

A
A numeric vector of discretized exposure vector (e.g., from a design matrix whose effect on expression values is of interest.

a
The numeric value indicating levels of A above against which comparisons are to be made.

subj_ids
A numeric vector of subject IDs to be passed directly to tmle::tmle when there are repeated measures; measurements on the same subject should have the exact same numerical identifier. These values will be coerced to numeric if not provided in the appropriate form (e.g., as character). The call to tmle::tmle will utilized a corrected version of the variance estimate from the efficient influence function.

family
(character) - specification of error family: "binomial" or "gaussian"

g_lib
(char vector) - library of learning algorithms to be used in fitting the propensity score E[A | W] (the nuisance parameter denoted "g" in the literature on targeted minimum loss-based estimation).

Q_lib
(char vector) - library of learning algorithms to be used in fitting the outcome regression E[Y | A, W] (the nuisance parameter denoted "Q" in the literature on targeted minimum loss-based estimation).

...
Additional arguments to be passed directly to tmle::tmle in fitting the targeted minimum loss-based estimator of the average treatment effect. Consult the documentation of that function for details.

Value

TMLE-based estimate of the relationship between biomarker expression and changes in an exposure variable, computed iteratively and saved in the tmleOut slot in a biotmle object.