Biomarker Evaluation with Targeted Minimum Loss Estimation of the ATE

Computes the causal target parameter defined as the difference between the biomarker expression values under treatment and those same values under no treatment, using Targeted Minimum Loss Estimation.

biomarkertmle(
  se,
  varInt,
  normalized = TRUE,
  ngscounts = FALSE,
  parallel = TRUE,
  bppar_type = NULL,
  future_param = NULL,
  bppar_debug = FALSE,
  cv_folds = 1,
  g_lib = c("SL.mean", "SL.glm", "SL.bayesglm", "SL.glmnet"),
  Q_lib = c("SL.mean", "SL.bayesglm", "SL.earth", "SL.xgboost"),
  ...
)

Arguments

se
A SummarizedExperiment containing microarray expression or next-generation sequencing data in the assays slot and a matrix of phenotype-level data in the colData slot.

varInt
A numeric indicating the column of the design matrix corresponding to the treatment or outcome of interest (in the colData slot of the SummarizedExperiment argument "se").

normalized
A logical indicating whether the data included in the assay slot of the input SummarizedExperiment object has been normalized externally. The default is set to TRUE with the expectation that an appropriate normalization method has been applied. If set to FALSE, median normalization is performed for microarray data.

ngscounts
A logical indicating whether the data are counts generated from a next-generation sequencing experiment (e.g., RNA-seq). The default setting assumes continuous expression measures as generated by microarray platforms.

parallel
A logical of whether or not to use parallelization in the estimation procedure. Invoking parallelization happens through a combination of calls to future and BiocParallel. If this argument is set to TRUE, multiprocess is used, and, if FALSE, sequential is used, alongside bplapply. Other options for evaluation through futures may be invoked by setting the argument future_param.

bppar_type
A character specifying the type of parallelization backend to be invoked by BiocParallel. Consult the manual page for BiocParallelParam for possible types and descriptions on their appropriate uses. The default for this argument is NULL, which silently uses DoparParam.

future_param
A character specifying the parallelization type to be invoked when using futures for evaluation. For a list of available types, please consult the documentation for plan. The default setting (this argument set to NULL) silently invokes multiprocess.

bppar_debug
A logical indicating whether or not to rely upon pkgBiocParallel. Setting this argument to TRUE, replaces the call to bplapply by a call to lapply, which significantly reduces the overhead of debugging. Note that invoking this option overrides all other parallelization arguments.

cv_folds
A numeric scalar indicating how many folds to use in performing targeted minimum loss estimation. Cross-validated estimates have been demonstrated to allow relaxation of certain theoretical conditions and and accommodate the construction of more conservative variance estimates.

g_lib
A character vector specifying the library of machine learning algorithms for use in fitting the propensity score P(A = a | W).

Q_lib
A character vector specifying the library of machine learning algorithms for use in fitting the outcome regression E[Y | A,W].

...
Additional arguments to be passed to drtmle in computing the targeted minimum loss estimator of the average treatment effect.

Value

S4 object of class biotmle, inheriting from SummarizedExperiment, with additional slots tmleOut and call, among others, containing TML estimates of the ATE of exposure on biomarker expression.

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following object is masked from ‘package:matrixStats’:
#> 
#>     count

#> The following object is masked from ‘package:Biobase’:
#> 
#>     combine

#> The following objects are masked from ‘package:GenomicRanges’:
#> 
#>     intersect, setdiff, union

#> The following object is masked from ‘package:GenomeInfoDb’:
#> 
#>     intersect

#> The following objects are masked from ‘package:IRanges’:
#> 
#>     collapse, desc, intersect, setdiff, slice, union

#> The following objects are masked from ‘package:S4Vectors’:
#> 
#>     first, intersect, rename, setdiff, setequal, union

#> The following objects are masked from ‘package:BiocGenerics’:
#> 
#>     combine, intersect, setdiff, union

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

library(biotmleData)
library(SuperLearner)

#> Loading required package: nnls

#> Super Learner

#> Version: 2.0-26

#> Package created on 2019-10-27

library(SummarizedExperiment)
data(illuminaData)

colData(illuminaData) <- colData(illuminaData) %>%
  data.frame() %>%
  dplyr::mutate(age = as.numeric(age > median(age))) %>%
  DataFrame()
benz_idx <- which(names(colData(illuminaData)) %in% "benzene")

biomarkerTMLEout <- biomarkertmle(
  se = illuminaData[1:2, ],
  varInt = benz_idx,
  parallel = FALSE,
  g_lib = c("SL.mean", "SL.glm"),
  Q_lib = c("SL.bayesglm", "SL.glmnet")
)

#> Warning: Sequential evaluation over many probes may take a long time.

#> Loading required package: globals

#> Loading required package: future

#> 
#> Attaching package: ‘future’

#> The following object is masked from ‘package:SummarizedExperiment’:
#> 
#>     values

#> The following object is masked from ‘package:GenomicRanges’:
#> 
#>     values

#> The following object is masked from ‘package:IRanges’:
#> 
#>     values

#> The following object is masked from ‘package:S4Vectors’:
#> 
#>     values

#> Loading required package: foreach

#> Loading required package: iterators

#> Loading required package: nloptr

#> Loading required package: quadprog

#> Loading required namespace: arm

#> Loading required namespace: glmnet