Moderated Statistical Tests for Influence Functions

Performs variance shrinkage via the empirical Bayes procedure of LIMMA on the observed data after a transformation moving the data to influence function space, based on the average treatment effect parameter.

modtest_ic(biotmle, adjust = "BH", ...)

Arguments

biotmle
biotmle object as generated by biomarkertmle

adjust
the multiple testing correction to be applied to p-values that are generated from the moderated tests. The recommended (and default) method is that of Benjamini and Hochberg. See topTable for a list of appropriate methods.

...
Other arguments to be passed directly to limma::topTable.

Value

biotmle object containing output from limma::lmFit and limma::topTable

Examples

library(biotmleData)
library(SummarizedExperiment)
data(biomarkertmleOut)

limmaTMLEout <- modtest_ic(biotmle = biomarkerTMLEout)